feature: 라우트 구조 수정 및 인기 검색어 게시판 구현
