+++ src/app/(boards)/entertain/page.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
+import { Board } from '@/views/boards';
 
 export default function Page() {
-  return <div>Entertainment</div>;
+  return <Board type="entertain" />;
 }
