+++ src/app/(boards)/entertain/post/[postId]/page.tsx
@@ -1,0 +1,6 @@
+import { Post } from '@/views/post';
+import React from 'react';
+
+export default function Page({ params }: { params: { postId: string } }) {
+  return <Post postId={params.postId} />;
+}
