+++ src/app/home/page.tsx
@@ -1,0 +1,6 @@
+import { Home } from '@/views/home';
+import React from 'react';
+
+export default function Page() {
+  return <Home />;
+}
