+++ src/views/boards/ui/index.ts
@@ -1,0 +1,3 @@
+import Board from './Board';
+
+export { Board };
