+++ src/views/hotBoards/ui/index.ts
@@ -1,0 +1,3 @@
+import HotBoard from './HotBoard';
+
+export { HotBoard };
