+++ src/views/post/ui/index.ts
@@ -1,0 +1,3 @@
+import Post from './Post';
+
+export { Post };
