--- src/widgets/hotBoard/ui/index.ts
@@ -1,7 +1,0 @@
-import AdvCarousel from './AdvCarousel';
-import CurrentHotRow from './CurrentHotRow';
-import MedalRow from './MedalRow';
-import { SortChip, SortChipItem } from './SortChip';
-import Fire from './icons/Fire';
-
-export { AdvCarousel, CurrentHotRow, MedalRow, SortChip, SortChipItem, Fire };
