+++ src/widgets/hotBoards/ui/AdvCarousel.tsx
