+++ src/widgets/hotBoards/ui/MedalRow.tsx
