+++ src/widgets/hotBoards/ui/icons/Fire.tsx
