+++ src/widgets/hotBoards/ui/icons/index.ts
