+++ src/widgets/hotBoards/ui/index.ts
@@ -1,0 +1,6 @@
+import AdvCarousel from './AdvCarousel';
+import HotBoardList from './HotBoardList';
+import { SortChip, SortChipItem } from './SortChip';
+import Fire from './icons/Fire';
+
+export { AdvCarousel, HotBoardList, SortChip, SortChipItem, Fire };
